+++ Duplicati/Library/Modules/Builtin/ResultSerialization/JsonFormatSerializer.cs
@@ -19,10 +19,8 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER 
 // DEALINGS IN THE SOFTWARE.
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
-using System.Reflection;
 using System.Threading.Tasks;
 using Duplicati.Library.Interface;
 using Newtonsoft.Json;
